+++ src/fonda/anomaly.cljs
@@ -1,4 +1,0 @@
-(ns fonda.anomaly)
-
-(defn anomaly? [m]
-  (some? (get m :cognitect.anomalies/anomaly)))